Fixing crash on Android due to new fb SDK version
Related to issue : https://github.com/facebook/react-native-fbsdk/issues/310

We had the same issue this morning while trying to build for Android. Seems to be related to new facebook SDK version 4.17.0, and the dependency not strict enough.

There might be a better solution to integrate the new SDK version but this fix is good for a temporary solution.