Adds native-only props to all views
There seems to be some issues with the new version of the native SDK (which is installed automatically by
Gradle) and this package (for both versions 0.5 and 0.6). I don't like this fix, since every native view that
I have programmed does not require this kind of setup. But it is the only solution I found in order to be
able to continue developing my app.