defer initialization until first usage
all these modules are initialized on app startup, which not only adds time to the startup, but all also cause a network request to happen.
Because of legal reasons we're not allowed to have any potential privacy sensitive request to be made before given consent, so having these initializations deferred until first usage makes us pass this requirement.