Integrate latest FBSDK on iOS and JS (#541)
There have been some breaking changes in the native FBSDK this year which prevents us from using the latest version. I've decided to help the problem and integrate the latest changes with this pull request.

Currently I've integrated the iOS FBSDK changes. My app builds and runs and I was able to login and obtain an access token. I'm not finished with these integrations.

Here are the list of things I want to do:

- double check iOS FBSDK changes to make sure I've got everything
- integrate Android FBSDK similarly
- document all the changes I've made
- test all the affected features

After I've done with these items this PR should be ready for a real review and consideration to merge it and support the latest FBSDK on both platforms.

It would be great if you could try this out and confirm that it works for you too. Everything should work on iOS.

I've started my branch from the latest published version which is `0.9.0`. I did this on purpose because `master` is broken currently due to `RCTComponentEvent` still being unpublished.

Your feedback would be greatly appreciated. 🙂

**[Update-2019-June-14]** This PR has been repurposed into integrating the latest FBSDK on iOS and JS only. The android part has been done in PR #546 

Fixes #533